+++ projects/ui-framework/src/lib/chips/chip/chip.component.spec.ts
@@ -2,7 +2,7 @@ import { NO_ERRORS_SCHEMA, SimpleChange } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ChipComponent } from './chip.component';
-import { ChipType } from '../chip.enum';
+import { ChipType } from '../chips.enum';
 import { By } from '@angular/platform-browser';
 import { ColorService } from '../../services/color-service/color.service';
