+++ projects/ui-framework/src/lib/chips/chips.enum.ts
