+++ projects/ui-framework/src/public_api.ts
@@ -76,8 +76,8 @@ export { ChipModule } from './lib/chips/chip/chip.module';
 export {
   ChipComponent
 } from './lib/chips/chip/chip.component';
-export { ChipType } from './lib/chips/chip.enum';
-export { Chip } from './lib/chips/chip.interface';
+export { ChipType } from './lib/chips/chips.enum';
+export { Chip } from './lib/chips/chips.interface';
 // InfoStrip Module
 export { InfoStripModule } from './lib/buttons-indicators/info-strip/info-strip.module';
 export { InfoStripComponent } from './lib/buttons-indicators/info-strip/info-strip.component';
@@ -198,7 +198,7 @@ export {
 // Chip Input Module
 export { ChipInputModule } from './lib/chips/chip-input/chip-input.module';
 export { ChipInputComponent } from './lib/chips/chip-input/chip-input.component';
-export { ChipInputChange } from './lib/chips/chip-input/chip-input.interface';
+export { ChipInputChange } from './lib/chips/chips.interface';
 // Social Module
 export { SocialModule } from './lib/form-elements/social/social.module';
 export { Social } from './lib/form-elements/social/social.enum';
